+++ src/ffi_c.cpp
@@ -497,6 +497,9 @@ GDK_DEFINE_C_FUNCTION_3(GA_send_transaction, struct GA_session*, session, const 
     struct GA_auth_handler**, call,
     { *call = make_call(new ga::sdk::send_transaction_call(*session, *json_cast(transaction_details))); });
 
+GDK_DEFINE_C_FUNCTION_3(GA_sign_message, struct GA_session*, session, const GA_json*, details, struct GA_auth_handler**,
+    call, { *call = make_call(new ga::sdk::sign_message_call(*session, *json_cast(details))); });
+
 GDK_DEFINE_C_FUNCTION_3(GA_twofactor_change_limits, struct GA_session*, session, const GA_json*, limit_details,
     struct GA_auth_handler**, call,
     { *call = make_call(new ga::sdk::change_limits_call(*session, *json_cast(limit_details))); })